requirements.txt
autogen-agentchat
autogen-ext[azure]
python-dotenv
azure-identity
azure-storage-blob
openai
Flask>=2.0
protobuf>=5.29.3,<6.0
autogen>=0.2.0
autogen-core>=0.5.6
python-dateutil>=2.8.2
streamlit>=1.32.0
azure-ai-projects==1.0.0b6
flask-cors
streamlit
streamlit-chat
requests



